Crud

+++ db/migrate/20200229211202_create_bulk_discounts.rb
@@ -2,7 +2,7 @@ class CreateBulkDiscounts < ActiveRecord::Migration[5.1]
   def change
     create_table :bulk_discounts do |t|
       t.string :name
-      t.integer :percentage_off
+      t.float :percentage_off
       t.integer :required_quantity
     end
   end
